Apollo: fix navigation behavior for selecting an artist -> album
From navigating to any artist, then any item, the previous behaviour handles
the "back" button by skipping the artist view and returning directly to the
main view (list of artists).

This patch changes this, in that the "back" button works exactly the same as
the "up" button (arrow in the actionbar) for when you follow the route
mentioned above.

Change-Id: I73055bf79ec54a51af7a6026f5a0191819e8e2b7